requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
requests==2.31.0
google-generativeai==0.3.2
openai==1.3.7
Pillow==10.1.0
python-multipart==0.0.6
uvicorn==0.24.0
gunicorn==21.2.0
